moved header files visible to other compilation units to the include directory
