+++ example/msg.cpp
@@ -1,4 +1,5 @@
-#include <comm/inet.hpp>
+/* example code for testing outgoing connections */
+#include <sadfs/comm/inet.hpp>
 
 #include <arpa/inet.h>
 #include <array>
