+++ src/bootstrap/sadcd-bootstrap.cpp
@@ -6,7 +6,7 @@
 
 // sadfs-specific includes
 #include "util.hpp"
-#include <sadcd/defaults.hpp>
+#include <sadfs/sadcd/defaults.hpp>
 
 // standard includes
 #include <array>
