+++ src/bootstrap/sadmd-bootstrap.cpp
@@ -6,7 +6,7 @@
 
 // sadfs-specific includes
 #include "util.hpp"
-#include <sadmd/defaults.hpp>
+#include <sadfs/sadmd/defaults.hpp>
 
 // standard includes
 #include <array>
