+++ src/sadmd/CMakeLists.txt
@@ -1,0 +1,3 @@
+# src/sadmd/CMakeLists.txt
+add_executable (sadmd
+	sadmd.cpp)
